backend/requirements.txt
asgiref==3.11.1
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.17.1
opencv-python-headless==4.13.0.92
pytesseract==0.3.13
Pillow==10.0.0
python-dotenv==1.2.2
sqlparse==0.5.5
typing_extensions==4.15.0
